requirements.txt
anthropic>=0.40.0
slack-bolt>=1.18.0
boto3>=1.34.0
aioboto3>=13.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
feedparser>=6.0.0
python-dateutil>=2.9.0
tenacity>=8.2.0
structlog>=24.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
